+++ sandbox.controller/src/kz/greetgo/sandbox/controller/model/ClientAddr.java
@@ -1,7 +1,7 @@
 package kz.greetgo.sandbox.controller.model;
 
 public class ClientAddr {
-    public int client;
+    public Long client;
     public AddrType type;
     public String street, house, flat;
 }
