+++ sandbox.controller/src/kz/greetgo/sandbox/controller/model/ClientPhone.java
@@ -1,7 +1,7 @@
 package kz.greetgo.sandbox.controller.model;
 
 public class ClientPhone {
-    public int client;
+    public Long client;
     public String number;
     public PhoneType type;
 }
